backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi>=0.135.1",
    "pydantic-settings>=2.13.1",
    "sqlmodel>=0.0.37",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
